forEach Remediation and separated out sse-client into its own client module
