+++ src/views/time.ejs
@@ -12,6 +12,6 @@
     <p>Open the console to see server and local times sent to all clients.</p>
 
     <script src="/js/bootstrap.min.js"></script>
-    <script src="/javascripts/time.js"></script>
+    <script src="/javascripts/time.js" type="module"></script>
 </body>
 </html>
